Temporary hack to see if it builds on ghc-8.10.3 on darwin
It seems that optP--cpp option goes to both hsc2hs and cpphs
tools. If it's there then hsc2hs fails, otherwise cpphs.
I'm trying to understand what has changed in GHC and fix that
issue. But as a temporary workaround I've split the package
into several ones so only one tool used in each and it fixed
this issue.